% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAST.R
\name{CAST.test}
\alias{CAST.test}
\title{CAST (Cohort Allelic Sum Test)}
\usage{
CAST.test(Y, X, Z = NULL, correction = "Price", perm = 100,
  alpha = NULL, c = NULL)
}
\arguments{
\item{Y}{a numeric vector of phenotypes. Affected individuals are coded 1 and unaffected individuals are coded 0.}

\item{X}{a numeric matrix of genotypes (row: individual, column: variant). Genotypes are coded 0,1 or 2 corresponding to the number of minor alleles.}

\item{Z}{optional numeric matrix of covariates. See Details.}

\item{correction}{type of covariates adjustment. See Details.}

\item{perm}{number of permutations. If not NULL,a "standard permutation procedure" is performed to compute the significance. See Details.}

\item{alpha}{error level. If not NULL,an "adaptive permutation procedure" is performed to compute the significance. See Details.}

\item{c}{precision of the p-value. If not NULL,an "adaptive permutation procedure" is performed to compute the significance. See Details.}
}
\value{
\item{p.value}{the p-value obtained by the phenotype permutation procedure. }
\item{stat}{the test statistic. }
}
\description{
Performs the CAST test described by Morgenthaler and Thilly (2007).
}
\details{
\subsection{Permutation procedures}{
 Two types of permutations procedures can be defined in the function: the standard permutation procedure and the adaptive permutation procedure.
 
 In the standard permutation procedure, the user specifies, in the argument "perm", the number of permutations to be done. The p-value will be \eqn{(R+1)(B+1)}. With \eqn{R} the number of permutation statistics superior to the observed statistic and \eqn{B} the number of permutations.
 
 In the adaptive permutation procedure, the user specifies, in the argument "alpha", the significance to achieve after multiple testing correction. In the argument "c", the estimation precision of the p-value. In function of these two paremeters, the maximal number of permutations and the maximal number of success to achieve will be computed. If the maximal number of success is reached, the p-value will be \eqn{R/B}. If not, the maximal number of permutations will be used to compute the p-value \eqn{(R+1)(B+1)}.
}

\subsection{Adjustment for covariates}{
 The adjustment for covariates is based on the permutation procedure described by Epstein et al. (2012).
 
 This package incorporates code from the package R BiasedUrn for the Multivariate Fisher NonCentral Hypergeometric distribution. The modification resides in a different installation settings that enables the analysis of thousands of individuals,
}
}
\examples{
X=matrix(sample(c(0,1,2), prob=c(0.999,0.001,0),replace=TRUE, 50000), nrow=500)
m=apply(X, MARGIN=2, 'sum')
X=X[,-which(m==0), drop=FALSE]
Y=sample(c(0,1), replace=TRUE, 500)
Z=NULL

alpha=0.05
c=0.2

res=CAST.test(X=X, Y=Y, alpha=alpha, c=c); res

}
\references{
TODO : REF of CAST test

Che R, Jack JR, Motsinger-Reif AA, Brown CC (2014) An adaptive permutation approach for genome-wide association study: evaluation and recommendations for use. BioData Min 7:9 . doi: 10.1186/1756-0381-7-9

Epstein MP, Duncan R, Broadaway KA, et al (2012) Stratification Score Matching Improves Correction for Confounding by Population Stratification in Case-Control Association Studies. Genet Epidemiol 36:195-205 . doi: 10.1002/gepi.21611

Agner Fog (2015). BiasedUrn: Biased Urn Model Distributions. R package version 1.07. https://CRAN.R-project.org/package=BiasedUrn
}
