% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDoEstRareDensity.R
\name{PlotDoEstRareDensity}
\alias{PlotDoEstRareDensity}
\title{PlotDoEstRareDensity}
\usage{
PlotDoEstRareDensity(pheno, geno, position, genome.size, autosomal = TRUE,
  gender = NULL, plot.density = "density", plot.counts = TRUE,
  plot.legend = TRUE, col.case = "red", col.ctrl = "black",
  args.plot = list(), args.legend = list(), args.text = list())
}
\arguments{
\item{pheno}{a numeric vector of phenotypes. Affected individuals are coded 1 and unaffected individuals are coded 0.}

\item{geno}{a numeric matrix of genotypes (row: individual, column: variant). Genotypes are coded 0,1 or 2 corresponding to the number of minor alleles.}

\item{position}{a numeric vector of variant positions.}

\item{genome.size}{a numeric value corresponding to the length of the analyzed region.}

\item{autosomal}{boolean. If TRUE, autosomal chromosome; FALSE, X chromosome.}

\item{gender}{numeric vector. 1=male; 2=female.}

\item{plot.density}{By default, the density of rare variant positions is plotted (plot.density="density"). If you want the density function being multiplied by the weighted allele frequency, use  "burden-density".}

\item{plot.counts}{If TRUE, rare allele counts are indicated.}

\item{plot.legend}{If TRUE, the legend is added.}

\item{col.case}{color to use for cases.}

\item{col.ctrl}{color to use for controls.}

\item{args.plot}{a list of additional graphical parameters for the 'plot' function.}

\item{args.legend}{a list of additional graphical parameters for the 'legend' function.}

\item{args.text}{a list of additional graphical parameters for the 'text' function.}
}
\description{
Rare variant association test comparing position density functions and mutation counts between cases and controls.
}
\examples{
pheno=rep(c(0,1), 500)
geno=matrix(sample(c(0,1),prob=c(0.99,0.01) ,1000*30, replace=TRUE), ncol=30)
position=sample(1:500,30)
genome.size=500

# plot.density="burden-density"
PlotDoEstRareDensity(pheno, geno, position, genome.size, plot.legend=TRUE,
                    plot.density="burden-density",col.case="blue",
                    args.plot=list(lwd=2, xaxt="no", yaxt="no", main="Gene"),
                    args.legend=list(x="topleft", lwd=2),
                    args.text=list(cex=1))

# plot.density="density"
PlotDoEstRareDensity(pheno, geno, position, genome.size, plot.legend=TRUE,
                    plot.density="density",col.case="blue",
                    args.plot=list(lwd=2, xaxt="no", yaxt="no", main="Gene"),
                    args.legend=list(x="topleft", lwd=2),
                    args.text=list(cex=1))



}
